image-description-main/image-description-main/requirements.txt
absl-py==0.15.0
altair==5.1.2
annotated-types==0.6.0
asgiref==3.5.0
astunparse==1.6.3
async-generator==1.10
attrs==23.1.0
audioread==3.0.1
Automat==20.2.0
beautifulsoup4==4.12.2
blinker==1.7.0
blis==0.7.11
Bottleneck==1.3.7
cachetools==5.3.2
catalogue==2.0.10
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
chromedriver-binary==122.0.6171.0.0
click==8.0.1
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.3
constantly==15.1.0
cryptography==36.0.1
cssselect==1.1.0
cycler==0.10.0
cymem==2.0.8
dateparser==1.1.8
decorator==4.4.2
deprecation==2.1.0
distlib==0.3.4
distro==1.8.0
Django==4.0.2
easydict==1.11
fastai==1.0.60
fastcore==1.5.29
fastdownload==0.0.7
fastprogress==1.0.3
filelock==3.4.2
Flask==1.1.2
flatbuffers==1.12
fsspec==2023.10.0
gast==0.4.0
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.23.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googletrans==4.0.0rc1
grpcio==1.34.1
gunicorn==21.2.0
h11==0.9.0
h2==3.2.0
h5py==3.1.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.17.3
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
imageio==2.32.0
imageio-ffmpeg==0.4.9
imbalanced-learn==0.11.0
imblearn==0.0
img2pdf==0.4.4
importlib-metadata==6.8.0
imutils==0.5.4
incremental==21.3.0
invoice2data==0.4.4
itemadapter==0.4.0
itemloaders==1.0.4
itsdangerous==2.0.1
Jinja2==3.0.1
jmespath==0.10.0
joblib==1.3.2
JPype1==1.4.1
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
Keras==2.4.3
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
langcodes==3.3.0
lazy_loader==0.3
libclang==16.0.6
librosa==0.10.1
llvmlite==0.41.1
lxml==4.8.0
Markdown==3.5
markdown-it-py==3.0.0
MarkupSafe==2.0.1
matplotlib==3.3.4
mdurl==0.1.2
ml-dtypes==0.2.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.0.7
murmurhash==1.0.10
mysql-connector==2.2.9
networkx==3.2.1
numba==0.58.1
numexpr==2.8.7
numpy==1.19.5
nvidia-ml-py3==7.352.0
oauthlib==3.2.2
ocrmypdf==15.3.0
opencv-python==4.8.1.78
opt-einsum==3.3.0
outcome==1.2.0
packaging==23.2
pandas==1.4.1
parsel==1.6.0
pdf2text==1.0.0
pdfminer==20191125
pdfminer.six==20221105
pdftables-api==2.0.0
pikepdf==8.5.2
Pillow==8.1.2
platformdirs==3.11.0
pluggy==1.3.0
pooch==1.8.0
preshed==3.0.9
proglog==0.1.10
Protego==0.2.1
protobuf==3.20.3
pyarrow==14.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio==0.2.14
pycparser==2.21
pycryptodome==3.19.0
pydantic==2.5.1
pydantic_core==2.14.3
pydeck==0.8.1b0
PyDispatcher==2.0.5
pygame==2.5.2
Pygments==2.16.1
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2022.1
PyYAML==6.0.1
queuelib==1.6.2
referencing==0.30.2
regex==2023.10.3
reportlab==4.0.6
requests==2.27.1
requests-file==1.5.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==13.6.0
rpds-py==0.10.6
rsa==4.9
sacremoses==0.1.1
safetensors==0.4.1
scikit-build==0.17.6
scikit-image==0.22.0
scikit-learn==1.0.2
scipy==1.8.0
Scrapy==2.6.1
seaborn==0.11.2
selenium==4.3.0
service-identity==21.1.0
six==1.15.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.2.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.4.2
srsly==2.4.8
streamlit==1.28.1
sympy==1.12
tabula-py==2.8.2
tenacity==8.2.3
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.14.0
tensorflow-cpu==2.5.0
tensorflow-estimator==2.5.0
tensorflow-intel==2.14.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==1.1.0
thinc==8.2.1
threadpoolctl==3.1.0
tifffile==2023.12.9
tldextract==3.2.0
tokenizers==0.14.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.1.1
torchvision==0.16.1
tornado==6.3.3
tqdm==4.63.1
transformers==4.16.0
trio==0.21.0
trio-websocket==0.9.2
Twisted==22.1.0
twisted-iocpsupport==1.0.2
typer==0.9.0
typing-extensions==3.7.4.3
tzdata==2021.5
tzlocal==5.2
urllib3==1.26.8
validators==0.22.0
virtualenv==20.13.0
w3lib==1.22.0
wasabi==1.1.2
watchdog==3.0.0
weasel==0.3.4
Werkzeug==2.0.1
wordcloud==1.9.2
wrapt==1.12.1
wsproto==1.1.0
zipp==3.17.0
zope.interface==5.4.0